Fixes for SDC2 DML for T_LKP_BASE_USER

+++ Python ETL/crmintegration/3_operational_to_analytical/Model_DML_SCD2_PT.sql
@@ -342,7 +342,7 @@ as
 	source_table.team_name,
 	case
 		when source_table."group" = '' then -1
-		else cast(substring(split_part(source_table."group",'"',3),2,len(split_part(source_table."group",'"',3))-2) as int)
+		else cast(substring(split_part(source_table."group",'"',3),2,len(split_part(source_table."group",'"',3))-2) as bigint)
 	end opr_group,
 	split_part(source_table."group",'"',6) dsc_group,
 	target_base_user_responsible.cod_base_user cod_base_user_responsible,
